% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgm2.R
\name{bgm2}
\alias{bgm2}
\title{Bayesian function (not used for the moment)
\code{bgm2} fonction graphique}
\usage{
bgm2(tab_input, min_c, max_c, Er, CVr, CV_Cobs, CV_Iobs, n_proj = 0)
}
\arguments{
\item{tab_input}{mcmc list}

\item{n_proj}{nombre d'annee de projection}
}
\description{
Bayesian function (not used for the moment)
\code{bgm2} fonction graphique
}
\examples{

#Les données d'entrée
data(tableau_pa)

resultats_pa<-delta (tab=tableau_pa, esp="BOBO",list_param=c("mois","zone","type_pirogue","motorisation","zone2","saison","engin_peche2"), type_donnee="commercial", effort="auto", titre="PA", param_test=c("mois","zone","type_pirogue","motorisation","zone2","saison","engin_peche2"), espece_id_list='espece', var_eff_list= c("nb_jour_peche", "nb_sorties"), ope_id=c("presence","code_pays","annee", "mois", "zone", "zone2" , "saison", "type_pirogue" , "motorisation" , "engin_peche", "engin_peche2"), col_capture='captures', logtrans="auto", interactions="auto", facteur_flotille="engin_peche2", seuil=0.05)
pre_global_model_pa<-sqldf('select annee as Year,sum(i_ab) as C,avg("E.dens") as I from resultats_pa group by annee',drv='SQLite')
bgm2(pre_global_model_pa)
}
