% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgm_output_4.R
\name{bgm_output_4}
\alias{bgm_output_4}
\title{Bayesian function (not used for the moment)
\code{bgm_output_5} fonction graphique}
\usage{
bgm_output_4(mcmc, data, n_proj = 4, multiple = FALSE, title)
}
\arguments{
\item{mcmc}{mcmc liste venant de la fonction bgm}
}
\description{
Bayesian function (not used for the moment)
\code{bgm_output_5} fonction graphique
}
