% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captures_aggregees.R
\docType{data}
\name{captures_aggregees}
\alias{captures_aggregees}
\title{Tableau de capture et effort de la pêche industrielle / Guinée}
\format{
A data frame
\describe{
  \item{code_pays}{EEZ}
  \item{annee}{Année}
  \item{espece}{Nom de l'espèce capturée}
  \item{capture}{Capture aggrégée}
  ...
}
}
\source{
\url{CNSHB}
}
\usage{
captures_aggregees
}
\description{
Tableau de capture et effort de la pêche industrielle / Guinée
}
\keyword{datasets}
