% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couplage_delta.R
\name{couplage_delta}
\alias{couplage_delta}
\title{la fonction couplage_delta s'applique au cas où le sous_modèle glm_positif est gaussien}
\usage{
couplage_delta(
  glm_posit,
  glm_pres,
  tab,
  esp,
  type_donnee,
  effort,
  list_param,
  espece_id_list,
  var_eff_list,
  ope_id,
  col_capture,
  facteur_flotille,
  seuil,
  titre
)
}
\arguments{
\item{glm_posit}{: Sous-modele abondance}

\item{glm_pres}{: Sous-modele presence/absence}

\item{tab}{: les donnees auxquelles on ajuste le modele}

\item{esp}{: espece}

\item{facteur_flotille}{: facteur pour lequel il est interessant de tracer les series temporelles d'indices modelises. (Ex : Trouver une flotille de tuning)}

\item{data_posit:}{donnees d'abondances (présences seulement)}

\item{data_pres}{: donnees presence_absence}
}
\description{
la fonction couplage_delta s'applique au cas où le sous_modèle glm_positif est gaussien
}
\examples{
 #PA
 couplage_delta(glm_ia_pa, glm_pres_pa, tableau_pa, "BOBO", type_donnee="commercial", effort="auto", list_param, espece_id_list, var_eff_list, ope_id, col_capture, facteur_flotille="engin_peche2", seuil=0.05, titre="PA")
}
