% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_poor_indicators.R
\name{data_poor_indicators}
\alias{data_poor_indicators}
\title{Estimate some data poors indicators}
\usage{
data_poor_indicators(data_freq, Catch_IA = NA, L_inf, year = NA, plot = T)
}
\arguments{
\item{Catch_IA}{data with weight and size vectors}

\item{L_inf}{Asymptotic length}

\item{year}{Year of interest you want to study and for which will be calculated 'f'}

\item{plot}{if TRUE, plot all the years}
}
\description{
\code{data_poor_indicators}
}
