% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{delta}
\alias{delta}
\title{la fonction couplage_delta s'applique au cas où le sous_modèle glm_positif est gaussien}
\usage{
delta(
  tab,
  esp,
  list_param,
  type_donnee,
  effort,
  titre,
  param_test,
  espece_id_list,
  var_eff_list,
  ope_id,
  col_capture,
  logtrans,
  interactions,
  facteur_flotille,
  seuil
)
}
\arguments{
\item{tab}{: les donnees auxquelles on ajuste le modele}

\item{esp}{: espece}

\item{facteur_flotille}{: facteur pour lequel il est interessant de tracer les series temporelles d'indices modelises. (Ex : Trouver une flotille de tuning)}

\item{glm_posit}{: Sous-modele abondance}

\item{glm_pres}{: Sous-modele presence/absence}

\item{data_posit:}{donnees d'abondances (présences seulement)}

\item{data_pres}{: donnees presence_absence}
}
\description{
la fonction couplage_delta s'applique au cas où le sous_modèle glm_positif est gaussien
}
\examples{
 #PA
 delta (tab=tableau_pa, esp="BOBO",list_param, type_donnee="commercial", effort="auto", titre="PA", param_test=list_param, espece_id_list, var_eff_list, ope_id, col_capture, logtrans="auto", interactions="auto", facteur_flotille="engin_peche2", seuil=0.05)
}
