% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_glm.R
\name{delta_glm}
\alias{delta_glm}
\title{this function realise the Delta coupling using the 2 precedents GLMs, extract the year factor, calculate the AI and plots.}
\usage{
delta_glm(glm_pres, glm_abundance, title, type)
}
\arguments{
\item{glm_abundance}{: outputs from the function model_IAplus()}

\item{title}{: first part of the title for the plots}

\item{type}{: type of the data series. Ex : "SC", "PA", etc...}

\item{glm_pres_abs}{: outputs from the function model_pres_abs()}
}
\description{
this function realise the Delta coupling using the 2 precedents GLMs, extract the year factor, calculate the AI and plots.
}
\examples{
#SC
table_annee_final_SC <- extract_glm(glm_pres_sc,glm_ab_sc, titreSC)
}
