% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_GLM_IA.R
\name{delta_GLM_IA}
\alias{delta_GLM_IA}
\title{Cette fonction réalise le couplage Delta à partir des GLM pres/abs et abondance, extrait le facteur année et réalise 3 plots de l'IA déduit.}
\usage{
delta_GLM_IA(glm_pres, glm_ab, titre, type)
}
\arguments{
\item{glm_pres}{: sorties de la fonction delta_presabs()}

\item{glm_ab}{: sorties de la fonction delta_abondance()}

\item{titre}{: première partie des titres de graphiques. Ex : "IA de pêches scientifiques"}
}
\description{
Cette fonction réalise le couplage Delta à partir des GLM pres/abs et abondance, extrait le facteur année et réalise 3 plots de l'IA déduit.
}
\examples{
#SC
table_annee_final_SC <- extract_glm(glm_pres_sc,glm_ab_sc, titreSC)
}
