% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort_fox_calculation.R
\name{effort_fox_calculation}
\alias{effort_fox_calculation}
\title{this function calculate the fishing effort E, the corresponding Fox effort Efox ; standardised on the last year (= most recent year : mEfox = 1)}
\usage{
effort_fox_calculation(data, catch_table, k)
}
\arguments{
\item{data}{: the table with the mean IA serie}

\item{catch_table}{: table with catches and years (format :}

\item{k}{: K factor of the Fox effort calculation. Either : 1, 3 or 5}
}
\description{
this function calculate the fishing effort E, the corresponding Fox effort Efox ; standardised on the last year (= most recent year : mEfox = 1)
}
\examples{

}
