% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Efox_calculation.R
\name{Efox_calculation}
\alias{Efox_calculation}
\title{Cette fonction calcule l'effort de peche E, l'effort de Fox correspondant Efox ; relativise à la dernière année (= année la plus récente)}
\usage{
Efox_calculation(data, table_capture, k)
}
\arguments{
\item{data}{: le tableau avec IA}

\item{table_capture}{: table avec les captures et les années (format :}

\item{k}{: facteur k du calcul de l'effort de Fox. Au choix : 1, 3 ou 5}
}
\description{
Cette fonction calcule l'effort de peche E, l'effort de Fox correspondant Efox ; relativise à la dernière année (= année la plus récente)
}
\examples{

}
