% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evo_an.R
\name{evo_an}
\alias{evo_an}
\title{evolution de l'indice d'abondance moyen dans le temps}
\usage{
evo_an(tab, titre)
}
\arguments{
\item{tab}{tableau de données}

\item{titre}{titre du graphique}
}
\description{
evolution de l'indice d'abondance moyen dans le temps
}
\examples{
PA #
evo_an(tableau_pa_ab, "PA")
evo_an(tableau_pi_ab, "PI")
evo_an(tableau_sc_ab, "SC")
}
