% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evo_facto.R
\name{evo_facto}
\alias{evo_facto}
\title{evolution de l'abondance par parametre (fonction graphique)
\code{evo_facto} fonction graphique}
\usage{
evo_facto(tab, facteur, titre)
}
\arguments{
\item{tab}{tableau de données}

\item{facteur}{facteur}

\item{titre}{titre du graphique}
}
\description{
evolution de l'abondance par parametre (fonction graphique)
\code{evo_facto} fonction graphique
}
\examples{
lapply(param_pa, evo_facto, tab=tableau_pa_ab, "PA")
lapply(param_pi, evo_facto, tab=tableau_pi_ab, "PI")
lapply(param_sc, evo_facto, tab=tableau_sc_ab, "SC")
}
