% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.density.bivar.R
\name{f.density.bivar}
\alias{f.density.bivar}
\title{A faire}
\usage{
f.density.bivar(x, y, nlevels, nb.points)
}
\arguments{
\item{x}{: A faire}

\item{y}{: A faire}

\item{nb.points}{: A faire}

\item{nlevels:}{A faire}
}
\description{
A faire
}
\examples{

 f.density.bivar ()
}
