% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fox_effort_calculation.R
\name{fox_effort_calculation}
\alias{fox_effort_calculation}
\title{Calculate Fox fishing effort}
\usage{
fox_effort_calculation(data, catch_table, k)
}
\arguments{
\item{data}{the table with the mean IA serie}

\item{catch_table}{table with catches and years (format :}

\item{k}{the k factor of the Fox effort calculation. Either : 1, 3 or 5}
}
\description{
\code{fox_effort_calculation} calculate the fishing effort E, the corresponding Fox effort Efox ; standardised on the last year (= most recent year : mEfox = 1)
}
\examples{
data(data_IA)
data(captures_aggregees)
data_captures <- captures_aggregees \%>\%  group_by(year) \%>\%  summarise(Catch = sum(c))
fox_effort_calculation(data_IA, data_captures, k=3)
}
