% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalised_model.R
\name{generalised_model}
\alias{generalised_model}
\title{this function develop the generalised production model from IA and Efox series}
\usage{
generalised_model(
  table_Efox,
  graph_param,
  vect_ini = c(10, -0.5, 1),
  vect_lower = c(0.5, -1, 0.01),
  vect_upper = c(100, 10, 9),
  warning_control = FALSE,
  max_iteration = 50
)
}
\arguments{
\item{table_Efox}{: table with IA and Efox}

\item{graph_param}{: vector gathering the graphhics aestetics parameters (format : c(lengthEfox, title, upper_x, upper_y, upper_ybis))}

\item{vect_ini}{: values of initialisation of the parameters a, b and p}

\item{vect_lower}{: values of lower limits of the parameters a, b and p}

\item{vect_upper}{: values of upper limits of the parameters a, b and p}

\item{warning_control}{: if TRUE, the nls won't stop its process if too many warnings are made.}

\item{max_iteration}{: number of iteration that will be runned.}

\item{log}{: by default, fit the model under log transformation of the AI. Else, fit without the log transformation (no start/limit)}
}
\description{
this function develop the generalised production model from IA and Efox series
}
\examples{

}
