% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_ai_plus.R
\name{glm_ai_plus}
\alias{glm_ai_plus}
\title{this function select the best GLM model possible for the abundance data, based on an AIC selection process (formula_select = "auto"). It runs the GLMs, plot the residuals analysis graphs and return the GLM outputs.}
\usage{
glm_ai_plus(tab, parameters, formula_select)
}
\arguments{
\item{tab}{: table of abundance data, extracted from the output of table_pres_abs (presence = 1)}

\item{parameters}{: list of parameters to test}

\item{formula_select}{: if "auto", the function select which formula as the lowest AIC. Else, run the selected formula.}
}
\description{
this function select the best GLM model possible for the abundance data, based on an AIC selection process (formula_select = "auto"). It runs the GLMs, plot the residuals analysis graphs and return the GLM outputs.
}
\examples{
 #PA
 glm_ia_pa <- glm_ia(tableau_pa_ab, param_pa, logtrans="auto", interactions="auto")
#PI
glm_ia_pi <- glm_ia(tableau_pi_ab, param_pi, logtrans="auto", interactions="auto") # 60 000 lignes, 9 facteurs = trop lourd pour interactions2
#SC
glm_ia_sc <- glm_ia(tableau_sc_ab, param_sc, logtrans="auto", interactions="auto")
}
