% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_ia.R
\name{glm_ia}
\alias{glm_ia}
\title{la fonction glm_ia établit le modèle d'indices d'abondances : (sous-modèle positif dans un GLM delta). Une selection des parametre est effectuee sur critere statistique (stepAIC)}
\usage{
glm_ia(tab, parametres, logtrans, interactions)
}
\arguments{
\item{tab}{: les données auxquelles ajuster le modèle}

\item{parametres}{: liste des parametres à tester}

\item{logtrans}{: pour automatiser la logtransformation des donnees si elle apporte une amélioration ("auto") ou pour forcer la logtransformation ("force")}

\item{interactions}{: prendre en compte les interactions. ("auto") ou ("force")}
}
\description{
la fonction glm_ia établit le modèle d'indices d'abondances : (sous-modèle positif dans un GLM delta). Une selection des parametre est effectuee sur critere statistique (stepAIC)
}
\examples{
 #PA
 glm_ia_pa <- glm_ia(tableau_pa_ab, param_pa, logtrans="auto", interactions="auto")
#PI
glm_ia_pi <- glm_ia(tableau_pi_ab, param_pi, logtrans="auto", interactions="auto") # 60 000 lignes, 9 facteurs = trop lourd pour interactions2
#SC
glm_ia_sc <- glm_ia(tableau_sc_ab, param_sc, logtrans="auto", interactions="auto")
}
