% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_IAplus.R
\name{glm_IAplus}
\alias{glm_IAplus}
\title{la fonction glm_ia établit le modèle d'indices d'abondances : (sous-modèle positif dans un GLM delta). Une selection des parametre est effectuee sur critere statistique (stepAIC) ; ajout de l'argument "formule_select" : soit "auto" et la fonction choisit le modèle par opti stat, soit on lui donne la formule que l'on souhaite tester. Permet de finaliser.}
\usage{
glm_IAplus(tab, parametres, formule_select)
}
\arguments{
\item{tab}{: les données auxquelles ajuster le modèle}

\item{parametres}{: liste des parametres à tester}

\item{formule_select}{: permet la selection manuel de la formule à tester dans le glm / "auto" sinon}

\item{interactions}{: prendre en compte les interactions. ("auto") ou ("force")}
}
\description{
la fonction glm_ia établit le modèle d'indices d'abondances : (sous-modèle positif dans un GLM delta). Une selection des parametre est effectuee sur critere statistique (stepAIC) ; ajout de l'argument "formule_select" : soit "auto" et la fonction choisit le modèle par opti stat, soit on lui donne la formule que l'on souhaite tester. Permet de finaliser.
}
\examples{
 #PA
 glm_ia_pa <- glm_ia(tableau_pa_ab, param_pa, logtrans="auto", interactions="auto")
#PI
glm_ia_pi <- glm_ia(tableau_pi_ab, param_pi, logtrans="auto", interactions="auto") # 60 000 lignes, 9 facteurs = trop lourd pour interactions2
#SC
glm_ia_sc <- glm_ia(tableau_sc_ab, param_sc, logtrans="auto", interactions="auto")
}
