% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_pres_abs.R
\name{glm_pres_abs}
\alias{glm_pres_abs}
\title{this function select the best GLM model possible for the presence/absence data, based on an AIC selection process (formula_select = "auto"). It runs the GLMs, plot the residuals analysis graphs and return the GLM outputs.}
\usage{
glm_pres_abs(tab, parameters, formula_select)
}
\arguments{
\item{tab}{: table of presence and absence data, coming as the output of table_pres_abs}

\item{parameters}{: list of parameters to test}

\item{formula_select}{: if "auto", the function select which formula as the lowest AIC. Else, run the selected formula.}
}
\description{
this function select the best GLM model possible for the presence/absence data, based on an AIC selection process (formula_select = "auto"). It runs the GLMs, plot the residuals analysis graphs and return the GLM outputs.
}
\examples{
 glm_pres_pa <- glm_pres(tableau_pa_pres, param_pa, interactions="auto")
 glm_pres_pi <- glm_pres(tableau_pi_pres, param_pi) #150 000 lignes, 9 facteurs = trop lourd pour les interactions
tableau_sc_pres <- pres_abs(tableau_sc, "PSEUDOTOLITHUS ELONGATUS")
tab <- tableau_sc_pres
tab_old <- tab
tab <- na.omit(tab)
nrow(tab_old) ; nrow(tab)
param_sc <- param_use(tableau_sc_pres, list_param)
param_sc <- param_sc[-1] #####type_campagne pose probleme
parametres <- param_sc
}
