% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indice_ab_pres.R
\name{indice_ab_pres}
\alias{indice_ab_pres}
\title{indice_ab_pres
Cette fonction est la reunion de deux anciennes fonctions décrites ci-dessous. Elle créé les tableaux necessaires pour le GLM et delta_GLM.
Fonction 1 (indice_ab) créé la variable d'abondance sur laquelle seront ajustés les modèles. Pour les pêches scientifiques, il s'agit des captures par surface chalutee. Pour les pêches commerciales, il s'agit des cpue nominales. Pour les cpue, cette fonction automatise le choix de la variable d'effort nominal selon un critère de disponibilité des données. Le choix de la variable d'effort nominal pourra également être manuel. Elle retourne une ligne par operation ou station.
Fonction 2 (pres_abs) créé le tableau qui sélectionne une espèce et donne sa presence/absence par station ou operation en vue d'un delta GLM (sous modèle presence/absence)}
\usage{
indice_ab_pres(
  tab,
  type_donnee,
  effort,
  esp,
  list_param,
  espece_id_list,
  var_eff_list,
  ope_id,
  col_capture,
  seuil
)
}
\arguments{
\item{tab}{: Tableau avec les captures et paramètres associés}

\item{type_donnee}{: "scientifique" ou "commercial"}

\item{effort}{: "auto" pour un choix automatique de la variable effort ou choix personnel ex "duree_peche","nombre_operation","nb_jour_peche", "nb_sorties", "surface_chalutee" si pas besoin de recalculer la surface chalutee dans les donnees scientifiques}

\item{esp}{: yyy nom_taxonomique ou commun selon tableau, espece dont on veut tester la presence. ex : "BOBO" , "PSEUDOTOLITHUS ELONGATUS"}

\item{list_param}{liste des param_tres}

\item{espece_id_list}{liste des espèces}

\item{var_eff_list}{liste des données d'éffort}

\item{ope_id}{columne des opérations}

\item{col_capture}{: listes d'identifiants de colonnes (voir prep_GLM_tab.Rmd)}

\item{seuil}{: seuil minimum de frequence des modalites a garder par facteur}
}
\value{
La fonction incice_ab_pres() retourne un seul tableau : tableau_pres a patir duquel on peut filtrer les presences pour obtenir tableau_ab
}
\description{
indice_ab_pres
Cette fonction est la reunion de deux anciennes fonctions décrites ci-dessous. Elle créé les tableaux necessaires pour le GLM et delta_GLM.
Fonction 1 (indice_ab) créé la variable d'abondance sur laquelle seront ajustés les modèles. Pour les pêches scientifiques, il s'agit des captures par surface chalutee. Pour les pêches commerciales, il s'agit des cpue nominales. Pour les cpue, cette fonction automatise le choix de la variable d'effort nominal selon un critère de disponibilité des données. Le choix de la variable d'effort nominal pourra également être manuel. Elle retourne une ligne par operation ou station.
Fonction 2 (pres_abs) créé le tableau qui sélectionne une espèce et donne sa presence/absence par station ou operation en vue d'un delta GLM (sous modèle presence/absence)
}
\examples{
#Liste des colonnes  parametres categoriels (facteurs) à explorer
list_param <- c("mois","zone","type_pirogue","motorisation","zone2","saison","engin_peche2", #PECHE ARTISANALE
"licence" , "tjb2", "puissance2", "longueur2", "profondeur2","nationalite"  ,"sect_cod" , "tsect_cod" ,"intitule",  #PECHE INDUSTRIELLE
"strate", "profond_deb2", "code_campagne2" , "code_projet")
#Liste des noms de colonnes possible pour l'effort de pêche
 var_eff_list <- c("duree_peche", "nombre_operation","nb_jour_peche", "nb_sorties", "larg_ouverture", "total_capture", "duree", "vitesse_chalutage")

   # liste des identifiants possible pour la colonne capture
 col_capture <- c("captures", "capture_retenue", "total_capture", "nombre")
    # liste des identifiants possible une observation
  ope_id=c("presence",
        "code_pays", "code_projet", "code_campagne2", "code_station ", "date_chalutage",
        "longitude_deb", "latitude_deb",  "longitude_fin", "latitude_fin",
        "strates", "profond_deb2", "type_campagne", #donnees scientifiques
        "code_navire", "date_operatio", "numero_operation","longitude_debut", "latitude_debut",
        "profondeur2", "puissance2", "tjb2", "longueur2", "nationalite", "licence", # donnees PI
        "annee", "mois", "zone", "zone2" , "saison", "type_pirogue" , "motorisation" , "engin_peche", "engin_peche2") #donnees PA
PA #
data(tableau_pa)


tableau_pa_pres <- indice_ab_pres (tableau_pa, "commercial", "auto", esp="BOBO", list_param, "espece", var_eff_list, ope_id, col_capture,seuil=0.005)
tableau_pa_ab <-  filter(tableau_pa_pres, presence==1)
kable(tableau_pa_ab[sample(nrow(tableau_pa_ab), 5), ])
kable(tableau_pa_pres[sample(nrow(tableau_pa_pres), 5), ])
PI #

tableau_pi_pres <- indice_ab_pres (tableau_pi, "commercial", "duree_peche",  esp="PSEUDOTOLITHUS ELONGATUS", list_param,  espece_id_list, var_eff_list, ope_id, col_capture, seuil=0.05)
tableau_pi_ab <-  filter(tableau_pi_pres, presence==1)
SC Kg#
tableau_sc_pres <- indice_ab_pres (tableau_sc, "scientifique", "auto",  esp="PSEUDOTOLITHUS ELONGATUS", list_param,  espece_id_list, var_eff_list, ope_id, col_capture="total_capture", seuil=0.05)
tableau_sc_ab <-  filter(tableau_sc_pres, presence==1)
SC nombre#
tableau_sc_pres <- indice_ab_pres (tableau_sc, "scientifique", "auto",  esp="PSEUDOTOLITHUS ELONGATUS", list_param,  espece_id_list, var_eff_list, ope_id, col_capture="nombre", seuil=0.05)
tableau_sc_ab <-  filter(tableau_sc_pres, presence==1)
table(tableau_sc_pres$presence)
}
