% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_calculation.R
\name{interaction_calculation}
\alias{interaction_calculation}
\title{Recalculation of year effects in case of interaction}
\usage{
interaction_calculation(
  data_table,
  proportion_table,
  glm_type,
  tab,
  effort,
  esp,
  list_param,
  var_eff_list,
  espece_id,
  catch_col,
  limit,
  formula_select
)
}
\arguments{
\item{data_table}{: outputs table from either model_pres_abs() or model_ai_plus}

\item{proportion_table}{: table giving the respective proportion of each modality}

\item{glm_type}{: either "pres/abs" or "abundance"}

\item{tab}{: initial data used for the glm modeling}
}
\description{
\code{delta_glm} realize the delta-coupling using the 2 precedents GLMs outputs, extract the year factor, calculate the AI and display the plots.
}
\examples{
data(tableau_sc)
glm_pres_abs <- model_pres_abs(tableau_sc, esp="PSEUDOTOLITHUS ELONGATUS", effort="auto", title="SC", list_param=c("annee", "saison", "strate"), var_eff_list=c("surface_chalutee"), espece_id='nom_taxonomique', catch_col='total_capture', interactions = TRUE, limit=0.0001, formula_select = "presence ~ strate * annee + saison")
prof <- c("0-5m", "5-10m", "10-15m", "15-30m", ">30m")
proportion <- c(0.10, 0.15, 0.15, 0.30, 0.30)
repartition_prof <- as.data.frame(cbind(prof, proportion))
repartition_prof$proportion <- as.numeric(repartition_prof$proportion)
rownames(repartition_prof) <- repartition_prof$prof

glm_pres_sc_recalcule <- interaction_calculation(data_table=glm_pres_abs, proportion_table=repartition_prof, glm_type="pres/abs", tab=tableau_sc, effort, esp, list_param, var_eff_list, espece_id, catch_col, limit, formula_select="presence ~ strate*annee + saison")
delta_glm(glm_pres_abs, glm_abundance, title="Scientific campaign GIN BOBO", type = "SC")
}
