% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_Lorenzen.R
\name{M_Lorenzen}
\alias{M_Lorenzen}
\title{M_Lorenzen}
\usage{
M_Lorenzen(list_age, M_inf, K, t0)
}
\arguments{
\item{list_age}{list of age for which we want to estimate M}

\item{M_inf}{Asymptotic natural mortality}

\item{K}{Growth parameter}

\item{t0}{Growth parameter}
}
\description{
\code{M_Lorenzen} Estimate natural mortality accordingly to Lorenzen methods (1996, 2000)
}
\examples{
list_age <- seq(0,3, by = 0.25)
M_inf <- 0.2
K <- 0.35
t0 <- -0.24

res <- M_Lorenzen(list_age, M_inf, K, t0)

res
}
