% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_ai.R
\name{mean_ai}
\alias{mean_ai}
\title{Mean of AI series}
\usage{
mean_ai(
  data_IA,
  MOY = TRUE,
  vect_year_elim,
  type_ref,
  type_other,
  fish_power,
  title
)
}
\arguments{
\item{data_IA}{data table containing the year in the first column, and the different AI series on the following ones}

\item{MOY}{if TRUE, then mean_3years is run}

\item{vect_year_elim}{vector containing the years to delete. Else, empty vector c()}

\item{type_ref}{Select AI for standardisation.}

\item{type_other}{Select others AI}

\item{fish_power}{A vector containing technological creep for each AI in type_other. Not applied if equal zero.}
}
\description{
\code{mean_ai} gather the different AI series (/!\ MAX 5 SERIES) in one single table, standardise one mean AI series and plot several graphs
}
\examples{
data(data_IA_JABBA)
data_IA_test <- data_IA_JABBA \%>\%  dplyr:::select(c(annee, ScientificSurvey_SEN, Artisanal_SEN))
mean_ai(data_IA=data_IA_test, MOY=T, vect_year_elim=c(), type_ref = "ScientificSurvey_SEN", type_other = "Artisanal_SEN", fish_power = c(0.05), title = "1st Try")

}
