% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_IA.R
\name{mean_IA}
\alias{mean_IA}
\title{Cette fonction réunie les séries d'IA (/!\ MAX 5 SERIES) dans une seule table, standardise et plot plusieurs graphs en relation}
\usage{
mean_IA(nb_table, MOY, vect_year_elim, ...)
}
\arguments{
\item{nb_table}{: nb de table inclut dans "..." (=le nb de série d'IA différentes)}

\item{MOY}{: si "oui" vient appliquer la fonction Moy_3ans()}

\item{vect_year_elim}{: vecteur contenant les différentes années à éliminer. Sinon, vecteur vide c()}

\item{...}{: les différentes tables d'IA (max 5)}
}
\description{
Cette fonction réunie les séries d'IA (/!\ MAX 5 SERIES) dans une seule table, standardise et plot plusieurs graphs en relation
}
\examples{

}
