% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_shepherd.R
\name{method_shepherd}
\alias{method_shepherd}
\title{Shepherd method}
\usage{
method_shepherd(data_freq, K, L_inf, step_class, plot = FALSE)
}
\arguments{
\item{data_freq}{input dataset data_freqle}

\item{K}{range aof tested values for K}

\item{L_inf}{range of tested valyes for L_inf}

\item{step_class}{length step between two length class}

\item{plot}{FALSE by default. IF TRUE, print plot of S}

\item{title}{fraction of the title in the plots}
}
\description{
\code{method_shepherd} returns an estimation of growth parameters K and L_inf (growth coefficient and growth asymptotic length) from Von Bertallanfy equation and the value of objective function.
}
\examples{
 data(data_freqleau_sc)
 model_ai_plus(data_freqleau_sc, esp = "PSEUDOTOLITHUS ELONGATUS", effort = "auto", title = "SC", list_param = c("annee", "saison", "strate"), var_eff_list=c("surface_chalutee"), espece_id='nom_taxonomique', catch_col='total_capture', interactions =FALSE, limit=0.001, formula_select = "log(i_ab+0.0001) ~ strate + annee + saison")

}
