% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moda_facto.R
\name{moda_facto}
\alias{moda_facto}
\title{nombre de donnees par modalite de chaque parametre}
\usage{
moda_facto(tab, facteur, titre)
}
\arguments{
\item{tab}{tableau de données}

\item{facteur}{facteurs}

\item{titre}{titre du graphique}
}
\description{
nombre de donnees par modalite de chaque parametre
}
\examples{
PA #
tableau_pa_pres <- indice_ab_pres (tableau_pa, "commercial", "auto", esp="BOBO", list_param,  espece_id_list, var_eff_list, ope_id, col_capture, seuil=0.005)
lapply(param_pa, moda_facto, tab=tableau_pa_ab, "PA") # peche artisanale
lapply(param_pi, moda_facto, tab=tableau_pi_ab, "PI") # peche industrielle
lapply(param_sc, moda_facto, tab=tableau_sc_ab, "SC") # campagnes scientifiques
}
