% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moda_few.R
\name{moda_few}
\alias{moda_few}
\title{nombre de donnees par modalite de chaque parametre}
\usage{
moda_few(tab, factor, seuil)
}
\arguments{
\item{tab}{tableau de données}

\item{seuil}{seuil}

\item{facteur}{facteurs}
}
\description{
nombre de donnees par modalite de chaque parametre
}
\examples{
tabz <- moda_few_rm(tableau_ab, names_facteur, 0.05)
}
