% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moda_few_rm.R
\name{moda_few_rm}
\alias{moda_few_rm}
\title{nombre de donnees par modalite de chaque parametre}
\usage{
moda_few_rm(tab, factor_list, seuil)
}
\arguments{
\item{tab}{tableau de données}

\item{seuil}{seuil}

\item{facteur_list}{facteurs}
}
\description{
nombre de donnees par modalite de chaque parametre
}
\examples{
tabz <- moda_few_rm(tableau_ab, names_facteur, 0.05)
}
