% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moda_removal.R
\name{moda_removal}
\alias{moda_removal}
\title{moda_removal}
\usage{
moda_removal(tab, factor, limit, ope_id2)
}
\arguments{
\item{tab}{: dataset table}

\item{factor}{: the factor from which we test each modality}

\item{limit}{: percentage representing the limit value under which the modality is removed}

\item{ope_id2}{: vector used in the join}
}
\description{
\code{moda_removal} removes the modality under-represented in the dataset, depending on the threshold value
}
