% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moda_selection.R
\name{moda_selection}
\alias{moda_selection}
\title{this function will select the modality that will be removed for each tested factor}
\usage{
moda_selection(tab, factor_list, limit, ope_id2)
}
\arguments{
\item{tab}{: dataset table}

\item{factor_list}{: list of the factor tested in the GLMs}

\item{limit}{: percentage representing the limit value under which the modality is removed}

\item{ope_id2}{: vector used in the join}
}
\description{
this function will select the modality that will be removed for each tested factor
}
\examples{
tabz <- moda_selection(tableau_ab, names_facteur, 0.05, ope_id2)
}
