% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_IAplus.R
\name{model_IAplus}
\alias{model_IAplus}
\title{Cette fonction récapitule la totalite des actions qui concernent le sous-modele abondance y compris la selection des variables}
\usage{
model_IAplus(
  tab,
  esp,
  param_test,
  type_donnee,
  effort,
  titre,
  list_param,
  espece_id_list,
  var_eff_list,
  col_capture,
  interactions,
  seuil,
  formule_select
)
}
\arguments{
\item{esp}{:  nom de l'espece comme precise dans les donnees}

\item{param_test}{: liste des facteurs potentiels à tester dans le glm}

\item{type_donnee}{: "scientifique" ou "commercial"}

\item{effort}{: "auto" pour un choix automatique de la variable effort ou choix personnel ex "duree_peche",   "nombre_operation","nb_jour_peche", "nb_sorties"}

\item{interactions}{: "Y" si on a une interaction dans notre GLM. "N" sinon.}

\item{formule_select}{: permet la selection manuel de la formule à tester dans le glm / "auto" sinon}

\item{tableau}{: tableau issu de la préparation des tableaux}
}
\description{
Cette fonction récapitule la totalite des actions qui concernent le sous-modele abondance y compris la selection des variables
}
\examples{
 #PA
 delta_abondance(tableau_pa, "BOBO", list_param, "commercial", "auto", titre="PA", list_param,  espece_id_list, var_eff_list, ope_id, col_capture, logtrans="auto", interactions="auto", seuil=0.05)
#PI
delta_abondance(tableau_pi, "PSEUDOTOLITHUS ELONGATUS", list_param, "commercial", "auto", titre="PI", list_param,  espece_id_list, var_eff_list, ope_id, col_capture, seuil=0.05)
#SC
delta_abondance(tableau_sc, "PSEUDOTOLITHUS ELONGATUS", list_param, "scientifique", "auto", titre="SC", list_param,  espece_id_list, var_eff_list, ope_id, col_capture, seuil=0.05)
}
