% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Moy_3ans.R
\name{Moy_3ans}
\alias{Moy_3ans}
\title{Cette fonction sert à "smooth" sur 3 ans (ou 2ans pour première et dernière ligne)}
\usage{
Moy_3ans(data)
}
\arguments{
\item{data}{: la table avec IA_MOYEN_STANDARD à lisser}
}
\description{
Cette fonction sert à "smooth" sur 3 ans (ou 2ans pour première et dernière ligne)
}
\examples{

}
