% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.panel.cor.R
\name{panel.cor}
\alias{panel.cor}
\title{Bayesian function (not used for the moment)
\code{panel.cor} fonction graphique}
\usage{
panel.cor(x, y, digits = 2, prefix = "", cex.cor)
}
\arguments{
\item{x}{: A faire}

\item{y}{: A faire}

\item{prefix}{: A faire}

\item{digit:}{A faire}

\item{cex.cor:}{A faire}
}
\description{
Bayesian function (not used for the moment)
\code{panel.cor} fonction graphique
}
\examples{

 f.density.bivar ()
}
