% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.panel.dens.R
\name{panel.dens}
\alias{panel.dens}
\title{Bayesian function (not used for the moment)
\code{panel.dens} fonction graphique}
\usage{
panel.dens(x, ...)
}
\arguments{
\item{x}{: A faire}
}
\description{
Bayesian function (not used for the moment)
\code{panel.dens} fonction graphique
}
\examples{

 panel.dens()
}
