% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_use.R
\name{param_use}
\alias{param_use}
\title{param_use
\code{param_use} nombre de donnees par modalite de chaque parametre}
\usage{
param_use(tab, list)
}
\arguments{
\item{tab}{tableau de données}

\item{list}{facteurs}
}
\description{
param_use
\code{param_use} nombre de donnees par modalite de chaque parametre
}
