% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_rectif_F.R
\name{pseudo_rectif_F}
\alias{pseudo_rectif_F}
\title{pseudo_rectif_F}
\usage{
pseudo_rectif_F(Rinit, FT, Mat_Cage, Mat_C, Mat_R, Mat_E, Mat_M)
}
\arguments{
\item{init}{Initial value for pseudo cohort analysis, equivalent to the F Terminal}
}
\description{
\code{pseudo_rectif_F} rectified VPA from a initial value of F Terminal
}
\examples{
 data(tableau_sc)
 model_ai_plus(tableau_sc, esp = "PSEUDOTOLITHUS ELONGATUS", effort = "auto", title = "SC", list_param = c("annee", "saison", "strate"), var_eff_list=c("surface_chalutee"), espece_id='nom_taxonomique', catch_col='total_capture', interactions =FALSE, limit=0.001, formula_select = "log(i_ab+0.0001) ~ strate + annee + saison")

}
