% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_estimate.R
\name{R_estimate}
\alias{R_estimate}
\title{R_estimate}
\usage{
R_estimate(FT, age, Mat_C, Mat_R, Mat_E, Mat_M)
}
\arguments{
\item{FT}{Terminal F}

\item{Mat_C}{data frame of catch, each column is a different gear}

\item{Mat_R}{vector of recruitment, if unknown, considered constant}

\item{Mat_E}{data frame of effort, each column is a different gear}

\item{Mat_M}{vector of mortality biological parameter}

\item{Rinit}{Initial value for pseudo cohort analysis, equivalent to the last year recruitment}
}
\description{
\code{R_estimate} returns Initial recruitment estimate
}
