% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_pseudo_rectif.R
\name{R_pseudo_rectif}
\alias{R_pseudo_rectif}
\title{R_pseudo_rectif}
\usage{
R_pseudo_rectif(Rinit, Mat_C, Mat_R, Mat_E, Mat_M)
}
\arguments{
\item{Rinit}{Initial value for pseudo cohort analysis, equivalent to the last year recruitment}

\item{Mat_C}{data frame of catch, each column is a different gear}

\item{Mat_R}{vector of recruitment, if unknown, considered constant}

\item{Mat_E}{data frame of effort, each column is a different gear}

\item{Mat_M}{vector of mortality biological parameter}
}
\description{
\code{R_pseudo_rectif} returns abundance, capturability and fishing mortality of a class age (F and N) from initial recruitment - readapted from Chassot et al. 2008
}
\examples{
#exemple tiré de Bertignac, 1987
espece <- "Dicentrarchus labrax"
age <- 18
Mat_C <- data.frame(catchA = c(8, 438, 13955, 16933, 47864, 22925,23117,4684,15853,10942,8839,18524,8141,9850,4485,3608,2642,1310), catchB = c(0, 184, 1827, 2245, 7158, 4792,7151,1767,5378,3524,1429,2533,1481,2470,1161,563,617,477), catchC = c(8, 1103, 11285, 5938, 9773, 2749,1722,228,678,483,368,213,87,265,224,87,85,0), catchD = c(7, 6435, 20817, 4023, 3656, 905,570,117,131,60,115,42,111,157,117,0,51,0))
Mat_Cage <- apply(Mat_C, 1, sum) # catch sum by age over fleets
Mat_M <- rep(0.2, age)
Mat_E <- data.frame(effortA = c(3500, 2900, 2200, 1500, 800, 800,800,800,800,800,800,800,800,800,800,800,800,800), effortB = c(800, 600, 450, 350, 200, 50,50,0, 0, 0, 0, 0, 0, 0, 0, 0, 0,0), efforC = c(550, 400, 200, 100, 100, 100,100,100,100,100,100,100,100,100,100,100,100,100), effortD = c(430, 600, 760, 900, 900, 1000,1100,1150,1250,1350,1350,1350,1350,1350,1350,1350,1350,1350))
Mat_R <- rep(100000, 18) # si indisponible, on suppose recrutement constant
Mat_M <- rep(0.2, 18)
Rinit <- 1.18*10^6#  Mat_N[1] # valeur issue de Bertignac, 1987
pseudo_rectif_R(Mat_Cage, Mat_C, Mat_R, Mat_E, Mat_M, Rinit)
}
