% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_pres_abs.R
\name{table_pres_abs}
\alias{table_pres_abs}
\title{Creating the table used for the GLMs}
\usage{
table_pres_abs(tab, esp, list_param, var_eff_list, espece_id, catch_col, limit)
}
\arguments{
\item{tab}{Table with the captures, fishing efforts and associated parameters}

\item{esp}{exact name of the studied species}

\item{list_param}{list of the tested parameters}

\item{var_eff_list}{list of the possible fishing effort column}

\item{espece_id}{exact name of the column indicating the species}

\item{catch_col}{exact name of the column indicating the catches}

\item{limit}{percentage representing the limit value under which the modality is removed}
}
\value{
the function return a table from which we can select only the \emph{presence = 1} to get the abundance table
}
\description{
\code{table_pres_abs} returns the table used as an argument in the glm modeling.
It creates the \emph{i_ab} and the \emph{presence} columns, and apply the modality selection functions.
}
\examples{
data(tableau_sc)
table_pres_abs(tableau_sc, effort="auto", esp="PSEUDOTOLITHUS ELONGATUS", list_param=c("annee", "saison", "strate"), var_eff_list=c("surface_chalutee"), espece_id='nom_taxonomique', catch_col='total_capture', limit=0.0001)


}
