% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau_pa.R
\docType{data}
\name{tableau_pa}
\alias{tableau_pa}
\title{Tableau de capture et effort de la pêche artisansale / Guinée}
\format{
A data frame :
\describe{
  \item{code_pays}{EEZ}
  \item{annee}{Année}
  \item{mois}{mois}
  \item{zone}{zone de capture}
  \item{port_site}{Port de débarquement}
  \item{type_pirogue}{Type de pirogue}
  \item{motorisation}{Motorisation TRUE/FALSE}
  \item{engin_peche}{Engin utilisé}
  \item{espece}{espèce capturée}
  \item{captures}{Captures en Kg}
  \item{nb_sorties}{Nombre de sorties}
  \item{nb_jour_peche}{Nombre de jours de pêche}
  \item{zone2}{regroupement de zones}
  \item{saison}{Regroupement de mois}
  \item{engin_peche2}{Regroupement d'engins}
  ...
}
}
\source{
\url{CNSHB}
}
\usage{
tableau_pa
}
\description{
Tableau de capture et effort de la pêche artisansale / Guinée
}
\keyword{datasets}
