% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau_pi.R
\docType{data}
\name{tableau_pi}
\alias{tableau_pi}
\title{Tableau de capture et effort de la pêche industrielle / Guinée}
\format{
A data frame
\describe{
  \item{code_pays}{EEZ}
  \item{code_navire}{Code of the fishing boat}
  \item{annee}{Année}
  \item{mois}{mois}
  \item{jour}{jour de la capture}
  \item{numero_operation}{Numero de l'opération de pêche}
  \item{espece}{Nom de l'espèce capturée}
  \item{capture_retenue}{Capture retenue pour le débarquement}
  \item{capture_rejetee}{Capture rejetée}
  \item{licence}{Type de licence autorisant l'activité de pêche}
  \item{longitude_fin}{Longitude (décimale) du trait de chalut}
  \item{latitude_fin}{Latitude (décimale) du trait de chalut}
  \item{date_debut_maree}{Date de début de la marée}
  \item{date_fin_maree}{Date de fin de la marée}
  \item{engin}{Engin utilisé par le bateau}
  \item{duree_peche}{Durée du trait}
  \item{nombre_operation}{Nombre d'opération, en cas de regroupement}
  \item{sect_cod.x}{Code Rectangle statistique}
  \item{tsect_cod.x}{Type de rectangle statistique}
  \item{intitule}{intitulé du rectangle statistique}
  \item{f_area}{intitulé du rectangle statistique}
  \item{id}{Code d'un second rectangle stat}
  \item{sect_cod.y}{Intitulé d'un second rectangle stat}
  \item{tsect_cod.y}{Type du second rectangle stat}
  \item{tsect_cod.y}{Type du second rectangle stat}
  \item{tsect_cod.y}{Type du second rectangle stat}
  ...
}
}
\source{
\url{CNSHB}
}
\usage{
tableau_pi
}
\description{
Tableau de capture et effort de la pêche industrielle / Guinée
}
\keyword{datasets}
