% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau_sc.R
\docType{data}
\name{tableau_sc}
\alias{tableau_sc}
\title{Tableau de données des campagnes scientifiques / Guinée}
\format{
A data frame :
\describe{
  \item{code_pays}{EEZ}
  \item{code_projet}{Code du projet }
  \item{code_campagne}{Code de la campagne}
  \item{code_station}{Code du trait de chalut}
  \item{nom_taxonomique}{Nom taxonomique de l'espèce}
  ...
}
}
\source{
\url{CNSHB}
}
\usage{
tableau_sc
}
\description{
Tableau de données des campagnes scientifiques / Guinée
}
\keyword{datasets}
