% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_Elefan.R
\name{VB_Elefan}
\alias{VB_Elefan}
\title{VB_Elefan}
\usage{
VB_Elefan(data_freq, Linf_guess, step_class, MA = 5, season = F)
}
\arguments{
\item{data_freq}{data with weight and size vectors}

\item{step_class}{}

\item{MA}{}

\item{season}{}

\item{L_inf}{}
}
\description{
\code{VB_Elefan} estimates R_init
}
\examples{
 data(data_weight_size)
 model_weight(data_weight_size)

}
