% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_size.R
\name{weight_size}
\alias{weight_size}
\title{Weight-size equation}
\usage{
weight_size(t, p)
}
\arguments{
\item{t}{time vector}

\item{p}{vector with values}
}
\description{
\code{weight_size} to plot size-weight relationship
}
\examples{
 t <- 1
 p <- c(0.005,3.16)
 weight_size(t,p)
}
