% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgm_output_5.R
\name{bgm_output_5}
\alias{bgm_output_5}
\title{Bayesian global model, on encapsule un modèle global générique dans la fonction}
\usage{
bgm_output_5(mcmc, data, n_proj = 0, multiple = FALSE, title)
}
\arguments{
\item{mcmc}{mcmc liste venant de la fonction bgm}
}
\description{
Bayesian global model, on encapsule un modèle global générique dans la fonction
}
