% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_poor_indicators.R
\name{data_poor_indicators}
\alias{data_poor_indicators}
\title{data_poor_indicators}
\usage{
data_poor_indicators(Catch_IA, L_inf, year, plot = F)
}
\arguments{
\item{Catch_IA}{data with weight and size vectors}

\item{L_inf}{}

\item{year}{}

\item{plot}{}
}
\description{
\code{data_poor_indicators} estimates R_init
}
\examples{
 data(data_weight_size)
 model_weight(data_weight_size)

}
