% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_glm.R
\name{delta_glm}
\alias{delta_glm}
\title{Delta coupling of 2 GLMs}
\usage{
delta_glm(
  tab_pres,
  tab_ia,
  esp,
  title,
  list_param,
  var_eff_list,
  espece_id,
  catch_col,
  limit,
  formula_select_pres,
  formula_select_ia,
  repartition = 0,
  temporal_factor = 0,
  data_type
)
}
\arguments{
\item{tab_pres}{input dataset table used for pres/abs model}

\item{tab_ia}{input dataset table used for ia model}

\item{esp}{exact name of the studied species}

\item{title}{title will be displayed in the plots}

\item{list_param}{list of the tested parameters}

\item{var_eff_list}{list of the possible fishing effort column}

\item{espece_id}{exact name of the column indicating the species}

\item{catch_col}{exact name of the column indicating the catches}

\item{limit}{percentage representing the limit value under which the modality is removed}

\item{formula_select_pres}{specify the final prese/abs model}

\item{formula_select_ia}{specify the final abundance model}

\item{repartition}{Surface by region strata (can be absolute value)}

\item{data_type}{Specify the type of data for the combining of different data_type}

\item{temporal_facto}{If other than annual and on a smaller time step (seasonal, quarter...). That way biomass mean will be calculated in this time step for each year}
}
\description{
\code{delta_glm} realize the delta-coupling using the 2 precedents GLMs, extract the year factor, calculate the AI and display the plots.
}
\examples{

data(tableau_sc)
tableau_sc_GIN <- tableau_sc \%>\% mutate(bathymetrie = case_when(
  profond_deb >= 05  & profond_deb <=10 ~ "5-10m",
  profond_deb > 10 & profond_deb <=15 ~ "10-15m",
 profond_deb > 15 & profond_deb <= 30 ~ "15-30m"
))


repartition <- as.data.frame(cbind(bathymetrie = c("5-10m", "10-15m", "15-30m"),
                                  proportion = c(661.5, 1009.3, 2872.3)))

repartition$proportion <- as.numeric(repartition$proportion)

delta_IA <- delta_glm(tab_pres = tableau_sc_GIN, tab_ia = tableau_sc_GIN, esp="PSEUDOTOLITHUS ELONGATUS", title="1st Try",
                      list_param=c("annee", "bathymetrie"), var_eff_list=c("surface_chalutee"), espece_id='nom_taxonomique',
                      catch_col='total_capture', limit=0.0001,
                      formula_select_pres = "presence ~ annee + bathymetrie",
                      formula_select_ia = "log(i_ab) ~ annee + bathymetrie",
                      repartition = repartition,
                      data_type = "SC")

}
