% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fox_model.R
\name{fox_model}
\alias{fox_model}
\title{Fox production model}
\usage{
fox_model(
  table_Efox,
  graph_param,
  a_start = 5,
  b_start = 3,
  log = TRUE,
  lower_b = 0
)
}
\arguments{
\item{table_Efox}{table with IA and Efox}

\item{graph_param}{vector gathering some graphics parameters (format : c(mE_range, title, upper_x, upper_y, upper_ybis))}

\item{a_start}{Initial value for a}

\item{b_start}{Initial value for b}

\item{log}{by default, fit the model under log transformation of the AI.}
}
\description{
\code{fox_model} this function fits the Fox global model from IA and Efox series
}
\examples{
data(data_IA)
data(captures_aggregees)
data_captures <- captures_aggregees \%>\%  group_by(year) \%>\%  summarise(Catch = sum(c))
table_Efox <- fox_effort_calculation(data_IA, data_captures, k=3)
mE_range <- 1.5
title <- "Fox model - Thiof"
graph_param <- c(mE_range, title) #param de mise en forme du graph
fox_model(table_Efox, graph_param, log = T, a_start = 1, b_start = 100)
}
