% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fox_model.R
\name{Fox_model}
\alias{Fox_model}
\title{Cette fonction réalise le modele de Fox à partir d'un IA et d'Efox}
\usage{
Fox_model(table_Efox, param_graph)
}
\arguments{
\item{data}{: le tableau avec IA}

\item{table_capture}{: table avec les captures et les années (format :}

\item{k}{: facteur k du calcul de l'effort de Fox. Au choix : 1, 3 ou 5}
}
\description{
Cette fonction réalise le modele de Fox à partir d'un IA et d'Efox
}
\examples{

}
