% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fox_model.R
\name{fox_model}
\alias{fox_model}
\title{this function develop the Fox global model from IA and Efox series}
\usage{
fox_model(
  table_Efox,
  graph_param,
  a_start = 5,
  b_start = 3,
  logarithmic = TRUE
)
}
\arguments{
\item{table_Efox}{: table with IA and Efox}

\item{graph_param}{: vector gathering the graphhics aestetics parameters (format : c(lengthEfox, title, upper_x, upper_y, upper_ybis))}

\item{log}{: by default, fit the model under log transformation of the AI. Else, fit without the log transformation (no start/limit)}
}
\description{
this function develop the Fox global model from IA and Efox series
}
\examples{

}
