% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalised_model.R
\name{generalised_model}
\alias{generalised_model}
\title{generalised_model}
\usage{
generalised_model(
  table_Efox,
  graph_param,
  a = c(1e-04, 0.001, 0.01),
  b = c(1e-04, 0.001, 0.01),
  m = c(0.5, 0.7, 0.9),
  warning_control = FALSE,
  max_iteration = 50,
  IC = T
)
}
\arguments{
\item{table_Efox}{: table with IA and Efox}

\item{graph_param}{: vector gathering the graphhics aestetics parameters (format : c(lengthEfox, title, upper_x, upper_y, upper_ybis))}

\item{a}{: values of initialisation and range of the parameter a (lower, init, upper)}

\item{b}{: values of initialisation and range of the parameter b}

\item{m}{: values of initialisation and range of the parameter m}

\item{warning_control}{: if TRUE, the nls won't stop its process if too many warnings are made.}

\item{max_iteration}{: number of iteration that will be runned.}

\item{log}{: by default, fit the model under log transformation of the AI. Else, fit without the log transformation (no start/limit)}
}
\description{
\code{generalised_model} Generalised production model from IA and Efox series
}
