% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_ai_plus.R
\name{glm_ai_plus}
\alias{glm_ai_plus}
\title{Fit a Generalized Linear Models on abundance data}
\usage{
glm_ai_plus(tab, parameters, formula_select)
}
\arguments{
\item{tab}{table of abundance data, extracted from the output of table_pres_abs}

\item{parameters}{list of parameters to test}

\item{formula_select}{if "auto", the function select which formula as the lowest AIC. Else, run the selected formula.}
}
\value{
\code{glm_ai_plus} can either return the best GLM model based on AIC comparison, or return the outputs of the GLM based on the formula given in \emph{formula_select}
}
\description{
\code{glm_ai_plus} returns the best GLM model possible for the abundance data, based on an AIC selection process (formula_select = "auto"). It runs the GLMs, plot the residuals analysis graphs and return the GLM outputs.
}
\examples{
data(tableau_sc)
list_param <- c("annee", "strate", "saison")
table_ex <- table_pres_abs(tableau_sc, effort="auto", esp="PSEUDOTOLITHUS ELONGATUS", list_param=c("annee", "saison", "strate"), espece_id='nom_taxonomique', var_eff_list=c("surface_chalutee"), catch_col='total_capture', limit=0.0001)
table_ex_abundance <- filter(table_ex, presence==1)
param <- param_use(table_ex_abundance, list_param)
for (i in 1:length(param)){
table_ex_abundance[,param[i]] <- as.factor(table_ex_abundance[,param[i]])
table_ex_abundance[,param[i]] <- droplevels(table_ex_abundance[,param[i]])
contrasts(table_ex_abundance[,param[i]]) <- contr.sum(levels(table_ex_abundance[,param[i]]))
}
glm_ai_plus(table_ex_abundance, param, formula_select = "auto")
}
