% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_pres_abs.R
\name{glm_pres_abs}
\alias{glm_pres_abs}
\title{la fonction glm_pres}
\usage{
glm_pres_abs(tab, parametres, formule_select)
}
\arguments{
\item{tab}{: les données auxquelles ajuster le modèle}

\item{parametres}{: liste des parametres à tester}

\item{formule_select}{: permet la selection manuel de la formule à tester dans le glm / "auto" sinon}
}
\description{
la fonction glm_pres
}
\examples{
 glm_pres_pa <- glm_pres(tableau_pa_pres, param_pa, interactions="auto")
 glm_pres_pi <- glm_pres(tableau_pi_pres, param_pi) #150 000 lignes, 9 facteurs = trop lourd pour les interactions
tableau_sc_pres <- pres_abs(tableau_sc, "PSEUDOTOLITHUS ELONGATUS")
tab <- tableau_sc_pres
tab_old <- tab
tab <- na.omit(tab)
nrow(tab_old) ; nrow(tab)
param_sc <- param_use(tableau_sc_pres, list_param)
param_sc <- param_sc[-1] #####type_campagne pose probleme
parametres <- param_sc
}
