% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_age.R
\name{M_lorenzen}
\alias{M_lorenzen}
\title{M_lorenzen}
\usage{
M_lorenzen(list_age, M_inf, K, t0)
}
\description{
\code{M_lorenzen} function to be optimized by stats::optimize, hence gives an estimation by iteration of F minimum value
}
\examples{
list_age <- seq(1,3.75, by = 0.25)
M_inf <- 0.2
b <- 3.15
t0 <- -0.23
K <- 0.2
espece <- "T. est"
age_max = 5
title = paste0("Mortality at age - Lorenzen method - \n", espece)
graph_param <- c(age_max, title)
M_lorenzen(M_inf, b, t0, K, list_age, graph_param)
}
