% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_3years.R
\name{mean_3years}
\alias{mean_3years}
\title{this function smooth the value over 3 years (or 2 years at the start and the end of the serie)}
\usage{
mean_3years(data)
}
\arguments{
\item{data}{: the table in mean_IA to smooth}
}
\description{
this function smooth the value over 3 years (or 2 years at the start and the end of the serie)
}
\examples{

}
