% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_3years.R
\name{mean_3years}
\alias{mean_3years}
\title{Cette fonction sert à "smooth" sur 3 ans (ou 2ans pour première et dernière ligne)}
\usage{
mean_3years(data)
}
\arguments{
\item{data}{: la table avec IA_MOYEN_STANDARD à lisser}
}
\description{
Cette fonction sert à "smooth" sur 3 ans (ou 2ans pour première et dernière ligne)
}
\examples{

}
