% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_ai.R
\name{mean_ai}
\alias{mean_ai}
\title{This function gather the different AI series (/!\ MAX 5 SERIES) in one single table, standardise one mean AI serie and plot several graphs}
\usage{
mean_ai(nb_table, MOY, vect_year_elim, ...)
}
\arguments{
\item{nb_table}{: number of tables included in "..."}

\item{MOY}{: if "Y", then mean_3years is runned}

\item{vect_year_elim}{: vector containing the years to delete. Else, empty vector c()}

\item{...}{: the different AI tables (max 5)}
}
\description{
This function gather the different AI series (/!\ MAX 5 SERIES) in one single table, standardise one mean AI serie and plot several graphs
}
\examples{

}
