% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_shepherd.R
\name{method_shepherd}
\alias{method_shepherd}
\title{Shepherd method}
\usage{
method_shepherd(data_freq, K, L_inf, step_class, plot = FALSE)
}
\arguments{
\item{data_freq}{input dataset data_freqle}

\item{K}{range aof tested values for K}

\item{L_inf}{range of tested valyes for L_inf}

\item{step_class}{length step between two length class}

\item{plot}{FALSE by default. IF TRUE, print plot of S}

\item{title}{fraction of the title in the plots}
}
\description{
\code{method_shepherd} returns an estimation of growth parameters K and L_inf (growth coefficient and growth asymptotic length) from Von Bertallanfy equation and the value of objective function.
}
\examples{
data(length_frequency)
lmax <- 48 # choose lmax according to some bibliography
K <- seq(0.33, 0.40, 0.01)
L_inf <- seq(0.95*lmax, 1.05 * lmax, 0.01)
step_class <- 1
method_shepherd(data_freq, K, L_inf, step_class,  plot=F)
}
