% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moda_facto.R
\name{moda_facto}
\alias{moda_facto}
\title{moda_facto}
\usage{
moda_facto(tab, facteur, title)
}
\arguments{
\item{tab}{tableau de données}

\item{facteur}{facteurs}

\item{title}{title}
}
\description{
\code{moda_facto} représente les données disponibles par modalite de chaque variable
}
