% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moda_selection.R
\name{moda_selection}
\alias{moda_selection}
\title{moda_selection}
\usage{
moda_selection(tab, factor_list, limit, ope_id2)
}
\arguments{
\item{tab}{: dataset table}

\item{factor_list}{: list of the factor tested in the GLMs}

\item{limit}{: percentage representing the limit value under which the modality is removed}

\item{ope_id2}{: vector used in the join}
}
\description{
\code{moda_selection} select the modality that will be removed for each tested factor
}
