% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ai_plus.R
\name{model_ai_plus}
\alias{model_ai_plus}
\title{Abundance-GLM modeling process}
\usage{
model_ai_plus(
  tab,
  esp,
  title,
  list_param,
  var_eff_list,
  espece_id,
  catch_col,
  interactions = FALSE,
  limit,
  formula_select,
  plot = FALSE,
  summary = FALSE,
  force_interaction = FALSE
)
}
\arguments{
\item{tab}{input dataset table}

\item{esp}{exact name of the studied species}

\item{title}{fraction of the title in the plots}

\item{list_param}{list of the tested parameters}

\item{var_eff_list}{list of the possible fishing effort column}

\item{espece_id}{exact name of the column indicating the species}

\item{catch_col}{exact name of the column indicating the catches}

\item{interactions}{FALSE by default. IF TRUE, print the interactions plots}

\item{limit}{percentage representing the limit value under which the modality is removed}

\item{formula_select}{if "auto", the function select which formula as the lowest AIC. Else, run the selected formula.}

\item{plot}{FALSE by default. If TRUE, print the data presentation histograms}

\item{summary}{FALSE by default. If TRUE, print the summary of the selected GLM}

\item{force_interaction}{Factorial plan can be incomplete, thus Anova type III won't work. Allow to coerce an examination of interactions by using Anova type I.}
}
\description{
\code{model_ai_plus} runs the actions for the GLM modeling process on abundance data.
}
\examples{
 data(tableau_sc)
 model_ai_plus(tableau_sc, esp = "PSEUDOTOLITHUS ELONGATUS", effort = "auto", title = "SC", list_param = c("annee", "saison", "strate"), var_eff_list=c("surface_chalutee"), espece_id='nom_taxonomique', catch_col='total_capture', interactions =FALSE, limit=0.001, formula_select = "log(i_ab+0.0001) ~ strate + annee + saison")

}
