% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ai_plus.R
\name{model_ai_plus}
\alias{model_ai_plus}
\title{this function run every action necessary for the GLM modelisation on abundance data}
\usage{
model_ai_plus(
  tab,
  esp,
  effort,
  title,
  list_param,
  espece_id,
  var_eff_list,
  catch_col,
  interactions,
  limit,
  formula_select
)
}
\arguments{
\item{tab}{: input dataset table}

\item{esp}{: exact name of the studied speciess}

\item{effort}{: "auto" for an automatic selection of the effort parameter or manual selection ex "duree_peche","nombre_operation","nb_jour_peche", "nb_sorties", "surface_chalutee"}

\item{title}{: fraction of the title in the plots}

\item{catch_col}{: exact name of the column indicating the catches}

\item{interactions}{: "Y" if there is an interaction in our GLM. "N" else}

\item{limit}{: percentage representing the limit value under which the modality is removed}

\item{formula_select}{: if "auto", the function select which formula as the lowest AIC. Else, run the selected formula.}
}
\description{
this function run every action necessary for the GLM modelisation on abundance data
}
\examples{
 #PA
 delta_abondance(tableau_pa, "BOBO", list_param, "commercial", "auto", titre="PA", list_param,  espece_id, var_eff_list, catch_col, logtrans="auto", interactions="auto", limit =0.05)
#PI
delta_abondance(tableau_pi, "PSEUDOTOLITHUS ELONGATUS", list_param, "commercial", "auto", titre="PI", list_param,  espece_id, var_eff_list, catch_col, limit =0.05)
#SC
delta_abondance(tableau_sc, "PSEUDOTOLITHUS ELONGATUS", list_param, "scientifique", "auto", titre="SC", list_param,  espece_id, var_eff_list, catch_col, limit =0.05)
}
