% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pres_abs.R
\name{model_pres_abs}
\alias{model_pres_abs}
\title{Presence/Absence-GLM modeling process}
\usage{
model_pres_abs(
  tab,
  esp,
  effort,
  title,
  list_param,
  var_eff_list,
  espece_id,
  catch_col,
  interactions,
  limit,
  formula_select
)
}
\arguments{
\item{tab}{: input dataset table}

\item{esp}{: exact name of the studied speciess}

\item{effort}{: "auto" for an automatic selection of the effort parameter or manual selection ex "duree_peche","nombre_operation","nb_jour_peche", "nb_sorties", "surface_chalutee"}

\item{title}{: fraction of the title in the plots}

\item{list_param}{: list of the tested parameters}

\item{var_eff_list}{: list of the possible fishing effort column}

\item{espece_id}{: exact name of the column indicating the species}

\item{catch_col}{: exact name of the column indicating the catches}

\item{interactions}{: "Y" if there is an interaction in our GLM. "N" else}

\item{limit}{: percentage representing the limit value under which the modality is removed}

\item{formula_select}{: if "auto", the function select which formula as the lowest AIC. Else, run the selected formula.}
}
\description{
\code{model_pres_abs} runs the actions for the GLM modeling process on presence/absence data.
}
\examples{
data(tableau_sc)
glm_pres <- model_pres_abs(tableau_sc, esp="PSEUDOTOLITHUS ELONGATUS", effort="auto", title="SC", list_param=c("annee", "saison", "strate"),  espece_id='nom_taxonomique', var_eff_list=c("surface_chalutee"), catch_col='total_capture', interactions = "N", limit=0.0001, formula_select = "presence ~ strate + annee + saison")
}
