% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pres_abs.R
\name{model_pres_abs}
\alias{model_pres_abs}
\title{la fonction delta_presabs récapitule les actions qui concernent le sous-modele presence-absence y compris la selection des variables}
\usage{
model_pres_abs(
  tab,
  esp,
  param_test,
  type_donnee,
  effort,
  titre,
  list_param,
  espece_id_list,
  var_eff_list,
  col_capture,
  interactions,
  seuil,
  formule_select
)
}
\arguments{
\item{esp}{= nom de l'espece comme precise dans les donnees}

\item{param_test}{= liste des facteurs potentiels à tester dans le glm}

\item{formule_select}{= permet la selection manuel de la formule à tester dans le glm / "auto" sinon}

\item{tableau}{= tableau issu de la préparation des tableaux}
}
\description{
la fonction delta_presabs récapitule les actions qui concernent le sous-modele presence-absence y compris la selection des variables
}
\examples{
 #PA
 delta_presabs(tableau_pa, "BOBO", list_param, "commercial", "auto", titre="PA", list_param,  espece_id_list, var_eff_list, ope_id, col_capture, interactions = "auto", seuil=0.05)
 #PI
 delta_presabs(tableau_pi, "PSEUDOTOLITHUS ELONGATUS", list_param, "commercial", "auto", titre="PI", list_param,  espece_id_list, var_eff_list, ope_id, col_capture, interactions = "auto", seuil=0.05)
 #SC
 delta_presabs(tableau_sc, "PSEUDOTOLITHUS ELONGATUS", list_param, "scientifique", "auto", titre="SC", list_param,  espece_id_list, var_eff_list, ope_id, col_capture, interactions = "auto", seuil=0.05)
}
