% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_weight.R
\name{model_weight}
\alias{model_weight}
\title{Weight size relationship and plot}
\usage{
model_weight(data_weight, a = 0.05, b = 2.5)
}
\arguments{
\item{data_weight}{data with weight and size vectors}

\item{a}{initialisation of a (set at 0.05)}

\item{b}{initialisation of b (set at 2.5)}
}
\description{
\code{model_weight} estimates R_init
}
\examples{
 data(data_weight_size)
 data_weight_size <- na.omit(data_weight_size)
 model_weight(data_weight_size)
}
