% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.panel.cor.R
\name{panel.cor}
\alias{panel.cor}
\title{A faire}
\usage{
panel.cor(x, y, digits = 2, prefix = "", cex.cor)
}
\arguments{
\item{x}{: A faire}

\item{y}{: A faire}

\item{prefix}{: A faire}

\item{digit:}{A faire}

\item{cex.cor:}{A faire}
}
\description{
A faire
}
\examples{

 f.density.bivar ()
}
