% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pres_facto.R
\name{pres_facto}
\alias{pres_facto}
\title{pres_facto}
\usage{
pres_facto(facteur, tab, titre)
}
\arguments{
\item{facteur}{facteurs}

\item{tab}{tableau de données}

\item{titre}{titre du graphique}
}
\description{
\code{pres_facto} Fonction to present observations for each factor
}
