% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_pres_abs.R
\name{table_pres_abs}
\alias{table_pres_abs}
\title{table_pres_abs}
\usage{
table_pres_abs(
  tab,
  effort,
  esp,
  list_param,
  espece_id,
  var_eff_list,
  catch_col,
  limit
)
}
\arguments{
\item{tab}{: Table with the captures, fishing efforts and associated parameters}

\item{effort}{: "auto" for an automatic selection of the effort parameter or manual selection ex "duree_peche","nombre_operation","nb_jour_peche", "nb_sorties", "surface_chalutee"}

\item{esp}{: exact name of the studied species}

\item{list_param}{: list of tested parameters}

\item{espece_id}{: exact name of the column indicating the species}

\item{var_eff_list}{: list of the possible fishing effort column}

\item{catch_col}{: exact name of the column indicating the catches}

\item{limit}{: percentage representing the limit value under which the modality is removed}
}
\value{
the function return a table from which we can select only the presence = 1 to get the abundance table
}
\description{
\code{table_pres_abs} returns the table used as an argument in the glm modeling.
It creates the i_ab and the presence columns, and apply the modalities selection functions.
}
\examples{
data(tableau_sc)
table_pres_abs(tableau_sc, effort="auto", esp="PSEUDOTOLITHUS ELONGATUS", list_param=c("annee", "saison", "strate"), espece_id='nom_taxonomique', var_eff_list=c("surface_chalutee"), catch_col='total_capture', limit=0.0001)


}
