% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_pres_abs.R
\name{table_pres_abs}
\alias{table_pres_abs}
\title{indice_ab_pres
this function create the table used as an argument in the glm modelisation. It creates the i_ab and the presence columns, apply the modality selection functions.}
\usage{
table_pres_abs(
  tab,
  effort,
  esp,
  list_param,
  espece_id,
  var_eff_list,
  catch_col,
  limit
)
}
\arguments{
\item{tab}{: Table with the captures, fishing efforts and associated parameters}

\item{effort}{: "auto" for an automatic selection of the effort parameter or manual selection ex "duree_peche","nombre_operation","nb_jour_peche", "nb_sorties", "surface_chalutee"}

\item{esp}{: exact name of the studied species}

\item{list_param}{: list of tested parameters}

\item{espece_id}{: exact name of the column indicating the species}

\item{var_eff_list}{: list of the possible fishing effort column}

\item{catch_col}{: exact name of the column indicating the catches}

\item{limit}{: percentage representing the limit value under which the modality is removed}
}
\value{
the function return a table from which we can select only the presence = 1 to get the abundance table
}
\description{
indice_ab_pres
this function create the table used as an argument in the glm modelisation. It creates the i_ab and the presence columns, apply the modality selection functions.
}
\examples{
a=1
}
