% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_recrut.R
\name{yield_recrut}
\alias{yield_recrut}
\title{yield_recrut}
\usage{
yield_recrut(
  a,
  b,
  Linf,
  K,
  t0,
  list_age,
  Mat_F,
  Mat_M,
  mf,
  F0.1 = F,
  Mat_W = NA
)
}
\arguments{
\item{a}{Initial value for pseudo cohort analysis, equivalent to the last year recruitment}

\item{b}{data frame of catch, each column is a different gear}

\item{Linf}{vector of recruitment, if unknown, considered constant}

\item{K}{data frame of effort, each column is a different gear}

\item{t0}{vector of mortality biological parameter}

\item{Mat_F}{}

\item{Mat_W}{}

\item{Mat_N2}{}

\item{Mat_Y}{}
}
\description{
\code{yield_recrut} function to be optimized by stats::optimize, hence gives an estimation by iteration of F minimum value
}
\examples{
-- IN PROCESS --
}
